ECS Service Discovery (#140)
* Implement AWS Discovery via ECS Service API

The initial implementation was done downstream in Liquidity
(https://github.com/dhpcs/liquidity).

The first approach was to use the conventions established by the
Ec2TagBasedSimpleServiceDiscovery implementation as a starting point.
The downstream commit on which that part of this contribution is based:
https://github.com/dhpcs/liquidity/commit/0513aacb154ae94b67e2f6ee54a2f2906d75c2c6.

That approach works, but by limiting ourselves to the old/stable AWS SDK
we are forced to accept blocking IO between our nodes and the AWS APIs.
To overcome this, I enhanced the initial implementation to use the
preview non-blocking AWS SDK.

The downstream commit on which that part of this contribution is based:
https://github.com/dhpcs/liquidity/commit/9dfbb701fba4e613b2efe4181494284f70a21bea.

In the downstream project the non-blocking version was developed
immediately after the blocking version - so the blocking version has not
actually been tested in anger. However, as can be seen in the downstream
commit, a main method was created and used to validate it works as a
unit (by doing runMain locally with an appropriate AWS config pointing
it to a real ECS service).

The non-blocking version was subject to the same test, but is also now
in use running Liquidity (albeit as a one-node-cluster most of the
time...).

The downstream commit in which the relevant configuration changes were
made:
https://github.com/dhpcs/liquidity/commit/1e376d25f02c4f8c0a17c2e3978660e0d2d87dec.

For me the non-blocking version is preferable (I'm happy using the
preview non-blocking AWS SDK), but I recognise it might not be desirable
to provide that as the only option in this contribution. Hence
including both, with appropriate namespace separations, so that other
downstream users can choose which they are most comfortable with.

In this commit I'm providing the implementation (which is the easy part,
because it was already done) in order to get a PR open sooner rather
than later. There's a TODO for documentation - that's the bit which
could take longer - hopefully within the next week or two though.

Note that the ECS service this was built for and is tested against is
using awsvpc mode networking*
(https://aws.amazon.com/about-aws/whats-new/2017/11/amazon-ecs-introduces-awsvpc-networking-mode-for-containers-to-support-full-networking-capabilities/).
awsvpc mode networking makes things a lot simpler by eliminating the
need for discovering ports too (as would be necessary with bridge mode),
while not also limiting containers to one per container instance (unlike
host mode). In other words, this solution is _not_ intended for use with
older ECS networking approaches; it's only for awsvpc mode.

*it's also using Fargate, but that shouldn't make a difference one way
or the other here.

* Create aws-api-ecs bootstrap-joining-demo

Includes:
1. CloudFormation template for creation of ECR repository, plus
   code-as-documentation launch script.
2. Minimal application.
3. Configuration of sbt-native-packager for generation of Docker image.
4. Script for publishing local image provided by (3) to ECR repository
   created by (1). We use docker:publishLocal and then tag and push the
   image manually in favour of just using docker:publish because this
   avoids the need to set dockerRepository within the SBT build (which
   would then require that we inject the AWS account ID into the build).
4. CloudFormation template for creation of ECS service (and dependencies),
   plus code-as-documentation launch script.

The security group includes a rule to allow public access to 19999 in
order that the formation of the cluster can be observed by running:

watch -n 1 -c "curl http://$taskInstanceIp:19999/cluster/members/ | python
-mjson.tool"

where $taskInstanceIp is the public IP of any of the task instances.

The remaining TODOs are:
1. Write the README for this addition.
2. Add to the README in the main docs.

* Add README for aws-api-ecs bootstrap-joining-demo

* Update to AWS ECS SDK 2.0.0-preview-9

* Complete discovery docs for AWS ECS Discovery

* Add missing Fargate, SBT, Docker and AWS CLI links

* Make tone consistent with general Akka documentation style

Also improve clarity of demo instructions and fix a typo.

* s/async-aws-api-ecs/aws-api-ecs-async/

* Do not create new AWS client for each call to lookup

* s/Console.out/System.err/

* Make note consistent with general Akka documentation style

* Make aws-api-ecs bootstrap-joining-demo part of main SBT build

If I'd looked beyond the EC2 demo I would have seen that the Kubernetes
and Marathon bootstrap-joining-demos are already doing this, and I
would have done it this way from the start. My initial thinking was a
bit too much of "be consistent with the EC2 patterns".

* Correctly Scalafmt AsyncEcsSimpleServiceDiscovery