Fixed aws-api ec2 use of tags
The withFilters method of AWS' DescribeInstancesRequest is always a set
operation when used with a collection. Ec2TagBasedSimpleServiceDiscovery
chains calls to withFilters and ends the chain by passing a collection
to the method, which effectively drops all filters previously set.
This change merges all ec2 filters in a single java list and sets the
filters of the DescribeInstancesRequest with the setFilters for better
clarity.