Make ECS clients lazily initialized
The commit message of
https://github.com/dhpcs/liquidity/commit/ad946d39c8a941914ce0825ef35e48c2dbecaa37
contains the full rationale, but in summary it's so that any failures in
creating the client (typically due to the AWS SDK provider chains not
finding credentials and/or region values) manifest as exceptions from
the call to lookup as opposed to the constructor. The reason this is
preferable is because of the way ServiceDiscovery itself instantiates
discovery implementations (with chained Try.recoverWith calls but
without ever logging failures).