Add a "Role-Based Access Control" heading
This will make the section stand out more and is easier to link to.

Also include the abbreviation "RBAC" in the text for SEO purposes. The abbreviation is used much more commonly than the fully spelled-out term.

Minikube now enables RBAC by default, as does Google Cloud, so this will be a common need for many users.