Use InetAddress type to denote IPv4 address
This is a refinement on #242 a part of a fix towards #187

This brings `apply(String, Option[Int])` back as a factory method for `ResolvedTarget`, and changes the type of `address` to `java.net.InetAddress`.

Since various orchestration engines just return a String for "host", I am doing a quick regex check to make sure it looks like an IPv4 address, and then parsing it into an
`InetAddress`.