Order addresses by address, then name
Workaround for #344

This might not have the desired effect when there are multiple routers with
different IP addresses doing name-based routing, and the service discovery
returns those with unique names but also addresses that point to those routers.
Perhaps we should cross that bridge when we know it appears in practice though.