By default assume akka-management.http.port for bootstrap
Refs #464, includes #502 which can be merged separately

Instead of discovering by port name 'management'.

This arguably makes discovery on Kubernetes a bit easier, since
the akka-management port does not need to be named any specific
way in the deployment .yml.

On the other hand, it makes the actual bootstrap/discovery process
a bit more complicated, and for our 'config discovery' example we'd
even have to add the 'port name' feature to that integration test
to be able to make it work at all.

I'm not convinced the convenience of not having to declare the port
name in the Kubernetes case weighs up to the extra complexity in the
bootstrap/discovery logic.