Remove pod-port-name option from kubernetes-api discovery (#502)
It does not make sense to configure a default at this level. It might make
sense to configure a default at the cluster-bootstrap level.

If no port name is specified in the query, return all ports - either for situations
where services always expose exactly 1 port, or perhaps when ports are
post-processed.