Don't return a port from kubernetes-api when none requested
Fixes #538

If no port is requested from kubernetes-api (because none is
configured), it should just return one target per pod, with no port set.
Cluster bootstrap already falls back to using the configured Akka HTTP
port if no port is returned from the lookup, so this will work.

I have tested this, and confirm it works (in particular, it works in an
Istio environment where the ports can't be specified in the container
spec).