Filter pods where the Akka container has 'waiting' status
We wait until pods become 'Running', but when using pods with multiple
containers, the pod becomes 'Running' as soon as the first primary
container starts.

https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/

In that case, when the Akka container fails to start, it also prevents
other pods (where the issue might be resolved) from starting. For that
reason, allow configuring which container is the Akka container, so we
can ignore it from discovery when it is 'waiting'.

Refs #890